requirements.txt
requests 
tqdm
ipaddress
maxminddb 
maxminddb-geolite2